feat(sync): recursive remote->local pull — the missing direction
Five Whys: infra's sovereign RAG sync could not use copia to PULL the index from
intel — `copia sync -r host:dir/ local/` errored because run_sync_recursive had
arms for local->remote and local->local only; remote->local hit a catch-all Err.
Recursive sync was built for push (GH-17); pull was never implemented. Root cause:
copia — the sovereign rsync replacement — was incomplete for pulls, forcing
consumers onto EXTERNAL rsync, defeating its whole purpose.

Fix: implement the pull direction, symmetric with push — discover_remote_files
(ssh `find -print0`, NUL-delimited) -> collect_dirs -> create_local_dirs ->
parallel STREAMING transfers (ssh cat piped to disk; never buffers the >1 GiB
index.json). The two directions are one algorithm: run_sync_dir(dir, ...) +
spawn_dir_transfers(dir, ...) + report_parallel_transfer(...).

Quality (never let defects ride): the copia bin file was already 1186 lines.
Split into 3 focused modules — main.rs (CLI), transfer.rs (SSH primitives),
dir_sync.rs (orchestration) — each now A- (99.8 / 99.3 / 92.0), all under the
size budget, clippy-clean, 247 + 3 tests + 10 doctests green. Provable contract
contracts/recursive-remote-pull-v1.yaml (pv-valid): structure preservation +
byte fidelity + parse relativization, 4 falsification tests.

Verified end-to-end: pulled a nested tree (sha256 3/3 match) AND the real 345 MB
sovereign-stack-index from intel at 226 MB/s; push re-verified. Bumps 0.1.4->0.1.5.
